Refactor/delete controller from replytable

+++ app/controllers/api/v1/reply_comments_controller.rb
@@ -1,21 +1,6 @@
 class Api::V1::ReplyCommentsController < ApplicationController
-  before_action :set_reply_comment, only: %i[ show update destroy ]
+  before_action :set_reply_comment, only: %i[update destroy]
   before_action :authenticate, only: %i[create update destroy]
-  
-  # create update deleteだけでいい。
-  # 表示するのは、postに含める
-
-  def index
-    # ReplyComment.create({user_id: User.first.id, post_id: Post.first.id, body: "コメントです"})
-    @reply_comments = ReplyComment.all
-
-    render json: {data: @reply_comments, message: "successfully get comments"},
-        status: 200
-  end
-
-  # def show
-  #   render json: @reply_comment
-  # end
 
   def create
     @reply_comment = ReplyComment.new(reply_comment_params)
@@ -29,8 +14,13 @@ class Api::V1::ReplyCommentsController < ApplicationController
     end
   end
 
-  # update, delete を自分のコメントだけ許可する
   def update
+    if @reply_comment.user_id != current_user.id
+      render json: {message: "更新する権限がありません。"},
+        status: 403
+      return
+    end
+
     if @reply_comment.update(reply_comment_params)
       render json: {data: @reply_comment, message: "successfully update comment"},
       status: 200
@@ -41,6 +31,12 @@ class Api::V1::ReplyCommentsController < ApplicationController
   end
 
   def destroy
+    if @reply_comment.user_id != current_user.id
+      render json: {message: "更新する権限がありません。"},
+        status: 403
+      return
+    end
+
     if @reply_comment.destroy
       render json: {data: @reply_comment, message: "コメントを削除しました。"},
         status: 200
@@ -56,6 +52,6 @@ class Api::V1::ReplyCommentsController < ApplicationController
     end
 
     def reply_comment_params
-      params.require(:reply_comment).permit(:user_id, :post_id, :body)
+      params.require(:reply_comment).permit(:post_id, :body).merge(user_id: current_user.id)
     end
 end
